test: move source location conversion test to owner module
Move the convert_base_loc regression test from build_hir into
source_location, where the helper is defined and shared.
