+++ compiler/crates/react_compiler_lowering/src/build_hir.rs
@@ -7346,42 +7346,3 @@ fn collect_identifier_node_ids_from_expr(
         _ => {}
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use react_compiler_ast::common::BaseNode;
-    use react_compiler_ast::common::Position as AstPosition;
-    use react_compiler_ast::common::SourceLocation as AstSourceLocation;
-
-    #[test]
-    fn convert_base_loc_keeps_start_end_offsets_separate_from_loc_index() {
-        let base = BaseNode {
-            node_type: Some("Identifier".to_string()),
-            start: Some(7),
-            end: Some(13),
-            loc: Some(AstSourceLocation {
-                start: AstPosition {
-                    line: 1,
-                    column: 2,
-                    index: Some(70),
-                },
-                end: AstPosition {
-                    line: 1,
-                    column: 8,
-                    index: Some(130),
-                },
-                filename: None,
-                identifier_name: None,
-            }),
-            ..Default::default()
-        };
-
-        let loc = convert_base_loc(&base).unwrap();
-
-        assert_eq!(loc.start.index, Some(70));
-        assert_eq!(loc.end.index, Some(130));
-        assert_eq!(loc.start_offset, Some(7));
-        assert_eq!(loc.end_offset, Some(13));
-    }
-}
